adding cache support to hibernate

+++ src/main/java/com/kaheshan/dummy/dao/PersonDao.java
@@ -9,6 +9,6 @@ public interface PersonDao {
    void save(Person person);
    void delete(int id);
    void add(Person person);
-   Person get(int id);
+   Person get(Long id);
    List<Person> listPersons();
 }
